# optimized GetAvailableUpdates in AddonManager