@pkerling 2019-05-22
Wouldn't it make more sense to investigate why `m_database.GetAddon` is (supposedly) so slow and optimize there?