@pkerling 2019-05-22
> Makes sense, not instead of this PR, but additionally, I'd say.

If you do a map lookup in `CAddoonDatabase` instead, it would make this PR completely obsolete (if that is easily possible)