@burlizzi 2019-05-23
m_database.GetAddon is not that much slow, it  just call SQL engine which is definitely slower than a std map search. As it is done in a loop the inefficiency is amplified.
the improvement here is to make a single SQL call at the beginning and iterate a local object