@burlizzi 2019-05-23
Database is in the name of CAddonDatabase, map lookup seems a violation of the design. 
I could think of some kind of caching but GetAddon is an atomic call, I wouldn't know how/when to populate the map cache without impacting on the existing code.